requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.1.0
discord.py @ git+https://github.com/Rapptz/discord.py@87bc79e6e311da60b024ae61a2a3757e55606aa6
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
numpy==1.23.1
python-dotenv==0.20.0
yarl==1.7.2
